consumer/requirements.txt
kafka-python
scikit-learn